requirements.txt
numpy>=1.21.0
Pillow>=9.0.0
PyMuPDF>=1.19.0
aiohttp>=3.8.0
 